Task B-12: add delete species, get species and breed by id, add breed…

+++ backend/src/PetHome.Application/Models/PagedList.cs
@@ -3,7 +3,7 @@
     public class PagedList<T>
     {
         public IReadOnlyList<T> Items { get; init; } = [];
-        public int TotalCount { get; init; }
+        public long TotalCount { get; init; }
         public int PageSize { get; init; }
         public int Page { get; init; }
         public bool HasNextPage => Page * PageSize < TotalCount;
